+++ info/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Add a helper function for enumerating Microsoft Office Trust Center security settings.
+- Add a check for Microsoft Office macro execution settings.
 
 ## 2026-01-28